pyproject.toml
[tool.poetry]
name = "pixify"
version = "0.1.0"
description = "An image messer-upper... now with AI!"
authors = ["Randall Sewell <randallsewell@fastmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.1.0"
flask-cors = "^5.0.0"
openai = "^1.58.1"
python-dotenv = "^1.0.1"
werkzeug = "^3.1.3"
load-dotenv = "^0.1.0"
gunicorn = "^23.0.0"
pillow = "^11.0.0"
requests = "^2.32.0"
python-jose = "^3.3.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
black = "^24.10.0"
flake8 = "^7.1.1"
pycparser = "^2.22"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
